# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AGRR Masters API
  version: 1.0.0
  description: |
    Public subset of the AGRR Masters JSON API for crop master data, agricultural tasks,
    task schedule blueprints, and external skill `setup_proposal` workflows.

    **Scopes (enforced per API key):**
    - `masters:read` — `GET` / `HEAD` on Masters endpoints, and `setup_proposal?mode=dry_run`
    - `masters:write` — create, update, delete, `setup_proposal?mode=apply`

    Session cookie authentication has full Masters access. API keys default to `masters:read` only on generate/regenerate.

    See [getting-started.md](./getting-started.md) for authentication and rate limits.
servers:
  - url: https://agrr.net
    description: Production
  - url: http://127.0.0.1:3000
    description: Local Docker (strangler-proxy)
security:
  - bearerApiKey: []
  - headerApiKey: []
paths:
...
  /api/v1/masters/crops/{crop_id}/crop_stages:
    parameters:
      - $ref: '#/components/parameters/CropId'
    get:
      tags: [crop_stages]
      summary: List crop stages
      operationId: listCropStages
      responses:
        '200':
          description: Stage array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CropStage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [crop_stages]
      summary: Create crop stage
      operationId: createCropStage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CropStageCreateRequest'
      responses:
        '201':
          description: Created stage
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /api/v1/masters/crops/{crop_id}/crop_stages/{id}:
    parameters:
      - $ref: '#/components/parameters/CropId'
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [crop_stages]
      summary: Show crop stage
      operationId: showCropStage
      responses:
        '200':
          description: Stage
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [crop_stages]
      summary: Update crop stage
      operationId: updateCropStage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CropStageUpdateRequest'
      responses:
        '200':
          description: Updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [crop_stages]
      summary: Delete crop stage
      operationId: destroyCropStage
      responses:
        '200':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      description: |
        User API key as Bearer token. Scopes are stored on the key (`masters:read`, `masters:write`).
        New keys default to `masters:read` only.
    headerApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        User API key header (alternative to Bearer). Same scope rules as bearerApiKey.
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: Browser session (UI only; not for server-to-server skills)
  parameters:
    CropId:
      name: crop_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found or not visible to the authenticated user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unprocessable:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
        error_code:
          type: string
        field_errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    CropStage:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        order:
          type: integer
    CropStageCreateRequest:
      type: object
      required: [crop_stage]
      properties:
        crop_stage:
          type: object
    CropStageUpdateRequest:
      type: object
      required: [crop_stage]
      properties:
        crop_stage:
          type: object
...